[CS EN] Properly test command failures
Since the script starts with `set -e`, any command (no matter where it
happens) that fails will fail the whole script. We need to be able to
test command statuses, which never happens if you run a command and then
try `if [[ -e $? 0 ]]` because the script has already exited. Moving
those commands directly to the `if` gets around this issue.

We also removed two instances of exit status checks that weren't
necessary.

This script can now wait for WildFly if it's run before WildFly is
ready.