Almost working SHA1 existing deployment check
The config.edn update changes the hash each time (even if the config is
this same, I think because it's changing the modification time of the
config.edn file stored in the .war file or some other similar metadata
kind of thing). I have a solution (I hope) for that coming soon.